+++ packages/moonstone/CHANGELOG.md
@@ -19,6 +19,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Picker` marquee issues with disabled buttons or Picker
 - `moonstone/Panels` to prevent loss of spotlight issue when moving between panels
 - `moonstone/Slider` by removing unnecessary repaints to the screen
+- `moonstone/VideoPlayer` to correctly position knob when interacting with media slider
 - `moonstone/VideoPlayer` to not read out the focused button when the media controls hide
 - `moonstone/MarqueeDecorator` to stop when unhovering a disabled component using `marqueeOn` `'focus'`